hopefully got zoom working for android

+++ ios/RCTNodeMediaClient/RCTNodeCameraManager.m
@@ -30,7 +30,7 @@ RCT_EXPORT_VIEW_PROPERTY(denoise, BOOL);
 RCT_EXPORT_VIEW_PROPERTY(smoothSkinLevel, int);
 RCT_EXPORT_VIEW_PROPERTY(cryptoKey, NSString)
 RCT_EXPORT_VIEW_PROPERTY(onChange, RCTBubblingEventBlock)
-RCT_EXPORT_VIEW_PROPERTY(zoomScale, int);
+RCT_EXPORT_VIEW_PROPERTY(zoomScale, float);
 
 
 RCT_EXPORT_METHOD(startprev:(nonnull NSNumber *)reactTag)
